ssl config: adjust to origin
When a transfer goes against another origin than the initial one,
do not add the following to the ssl configuration: client cert,
client key, srp user/pass, pinned key.
